(MODULES-10052) Parameterize WaitForExit for PXP agent
Change the timeout value to accept and define a parameter and a default value of 2 minutes. Default WaitForExit timeout value is causing puppet agent upgrades to fail on systems that are running longer than 2 minutes. This allows the user to set a desired timeout or accept the default value of 2 minutes.